app: move geo listener & tweak loader
